backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfplumber==0.10.3
python-docx==1.1.0
spacy==3.7.4
scikit-learn==1.4.1.post1
numpy>=1.26.0
supabase==2.4.0
python-dotenv==1.0.1
httpx>=0.24,<0.26
